Fix pointer int casting

+++ config.h
@@ -1,15 +1,15 @@
 #pragma once
 
+#include <stdint.h>
+
 #ifndef __UNCONST
-#define __UNCONST(a)   ((void *)(unsigned long)(const void *)(a))
+#define __UNCONST(a)   ((void *)(uintptr_t)(const void *)(a))
 #endif /* __UNCONST */
 
 #ifndef USE_ARG
 #define USE_ARG(x)       /*LINTED*/(void)&(x)
 #endif /* USE_ARG */
 
-
-#include <stdint.h>
 
 typedef uint8_t u_char;
